pyproject.toml
[tool.poetry]
name = "tango-extensions"
version = "0.0.0"                                               # specified by poetry-dynamic-versioning
description = "Extension modules for allenai/tango (ai2-tango)"
authors = ["Shunsuke KITADA <shunsuke.kitada.0831@gmail.com>"]
readme = "README.md"
packages = [{ include = "tango_ext" }]

[tool.poetry.dependencies]
python = "^3.9"
urllib3 = "<2"
ai2-tango = { extras = [
    "datasets",
    "examples",
    "torch",
    "transformers",
    "wandb",
], version = ">=1.0.0" }
torch = { version = "!=2.0.1", optional = true }
diffusers = { version = ">=0.18.0", optional = true }
accelerate = { version = "^0.22.0", optional = true }
albumentations = { version = ">=1.0.0", optional = true }

[tool.poetry.extras]
diffusers = ["diffusers", "accelerate", "torch"]
albumentations = ["albumentations"]
all = ["diffusers", "accelerate", "torch", "albumentations"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.285"
black = "^23.7.0"
mypy = "^1.5.1"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
